Consolidate local plugins into core modules
