+++ bom/pom.xml
@@ -32,23 +32,11 @@
         </dependency>
         <dependency>
             <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-filestore-local</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
             <artifactId>spals-appbuilder-message-core</artifactId>
         </dependency>
         <dependency>
             <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-message-local</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
             <artifactId>spals-appbuilder-nosqlstore-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-nosqlstore-local</artifactId>
         </dependency>
     </dependencies>
