+++ config/src/main/java/net/spals/appbuilder/config/ConsumerConfig.java
@@ -19,6 +19,7 @@ public interface ConsumerConfig {
     String ACTIVE_KEY = "active";
     String CHANNEL_KEY = "channel";
     String GLOBAL_ID_KEY = "globalId";
+    String FORMAT_KEY = "format";
     String SOURCE_KEY = "source";
 
     @JsonProperty(ConsumerConfig.ACTIVE_KEY)
@@ -31,6 +32,9 @@ public interface ConsumerConfig {
 
     @JsonProperty(ConsumerConfig.CHANNEL_KEY)
     String getChannel();
+
+    @JsonProperty(ConsumerConfig.FORMAT_KEY)
+    String getFormat();
 
     @JsonProperty(ConsumerConfig.GLOBAL_ID_KEY)
     String getGlobalId();
