+++ message-core/src/main/java/net/spals/appbuilder/message/core/consumer/MessageConsumerCallback.java
@@ -1,11 +1,13 @@
 package net.spals.appbuilder.message.core.consumer;
 
-import net.spals.appbuilder.message.core.model.Message;
+import net.spals.appbuilder.config.ConsumerConfig;
+
+import java.util.Map;
 
 /**
  * @author tkral
  */
 public interface MessageConsumerCallback {
 
-    void processMessage(Message message);
+    void processMessage(ConsumerConfig consumerConfig, Map<String, Object> payload);
 }
