+++ message-core/src/main/java/net/spals/appbuilder/message/core/consumer/MessageConsumerPlugin.java
@@ -1,13 +1,14 @@
 package net.spals.appbuilder.message.core.consumer;
 
 import net.spals.appbuilder.config.ConsumerConfig;
+import net.spals.appbuilder.message.core.formatter.MessageFormatter;
 
 /**
  * @author tkral
  */
 public interface MessageConsumerPlugin {
 
-    void start(ConsumerConfig consumerConfig);
+    void start(ConsumerConfig consumerConfig, MessageFormatter messageFormatter);
 
     void stop(ConsumerConfig consumerConfig);
 }
