--- message-core/src/main/java/net/spals/appbuilder/message/core/model/Message.java
@@ -1,22 +1,0 @@
-package net.spals.appbuilder.message.core.model;
-
-import org.inferred.freebuilder.FreeBuilder;
-
-import java.util.Map;
-
-/**
- * @author tkral
- */
-@FreeBuilder
-public interface Message {
-
-    Map<String, Object> getPayload();
-
-    String getChannel();
-
-    String getGlobalId();
-
-    String getTag();
-
-    class Builder extends Message_Builder {  }
-}
