+++ nosqlstore-core/pom.xml
@@ -43,6 +43,10 @@
             <artifactId>freebuilder</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.mapdb</groupId>
+            <artifactId>mapdb</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
         </dependency>
