+++ pom.xml
@@ -59,12 +59,9 @@
         <module>config</module>
         <module>config-test</module>
         <module>filestore-core</module>
-        <module>filestore-local</module>
         <module>message-core</module>
-        <module>message-local</module>
         <module>nosqlstore-core</module>
-        <module>nosqlstore-local</module>
-        <module>nosqlstore-local-test</module>
+        <module>nosqlstore-core-test</module>
     </modules>
 
     <dependencyManagement>
@@ -132,27 +129,12 @@
             </dependency>
             <dependency>
                 <groupId>net.spals.appbuilder</groupId>
-                <artifactId>spals-appbuilder-filestore-local</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>net.spals.appbuilder</groupId>
                 <artifactId>spals-appbuilder-message-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
                 <groupId>net.spals.appbuilder</groupId>
-                <artifactId>spals-appbuilder-message-local</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>net.spals.appbuilder</groupId>
                 <artifactId>spals-appbuilder-nosqlstore-core</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>net.spals.appbuilder</groupId>
-                <artifactId>spals-appbuilder-nosqlstore-local</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
